Create mxdirectxinfo
